+++ docs/changelog.md
@@ -11,6 +11,7 @@ nav_order: 2
 ### Bugfixes
 
 - Correctly retry sending a TunnellingRequest if no TunnellingAck was received for the first time for UDP tunnelling connections.
+- Properly log repeated heartbeat errors
 
 ## 1.0.0 Support for lukewarm temperatures 2022-08-13